requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-groq
langchain-core
faiss-cpu
unstructured
pypdf